requirements.txt
## Core HTTP/config/logging
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.1.8
httpcore==1.0.9
httpx==0.27.2
idna==3.11
packaging==25.0
python-dotenv==1.0.1
requests==2.32.3
sniffio==1.3.1
typing_extensions==4.13.2
urllib3==2.4.0

## Database layer
greenlet==3.2.4
pgvector==0.2.5
psycopg==3.2.1
psycopg-binary==3.2.1
psycopg2-binary==2.9.11
SQLAlchemy==2.0.36

## Embeddings / LLMs
filelock==3.19.1
fsspec==2025.9.0
huggingface-hub==0.36.0
regex==2025.11.3
safetensors==0.6.2
sentence-transformers==3.0.1
tokenizers==0.22.1
torch==2.9.1
transformers==4.57.1
triton==3.5.1

## NVIDIA CUDA runtime wheels
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90

## Numerics / ML utilities
joblib==1.5.2
networkx==3.5
numpy==1.26.4
scikit-learn==1.7.2
scipy==1.16.3
threadpoolctl==3.6.0

## Document parsing / scraping
beautifulsoup4==4.13.4
Markdown==3.10
markdown-it-py==4.0.0
mdurl==0.1.2
PyMuPDF==1.26.6
pymupdf4llm==0.2.0
selenium==4.33.0
soupsieve==2.8

## GUI (PyQt) and helpers
PyQt5==5.15.11
PyQt5-Qt5==5.15.18
PyQt5_sip==12.17.1
pillow==12.0.0
PySocks==1.7.1

## FastAPI/Starlette/UVicorn stack
fastapi==0.115.0
h11==0.16.0
httptools==0.7.1
starlette==0.38.6
uvicorn==0.30.6
uvloop==0.22.1
watchfiles==1.1.1

## CLI / formatting / misc
click==8.1.8
rich==13.9.2
shellingham==1.5.4
tabulate==0.9.0
typer==0.12.5

## Testing
iniconfig==2.3.0
pluggy==1.6.0
pytest==8.3.3
pytest-qt==4.5.0

## Math / rendering
Pygments==2.19.2
sympy==1.14.0

## Trio/websocket stack
outcome==1.3.0.post0
trio==0.30.0
trio-websocket==0.12.2
websockets==15.0.1
websocket-client==1.8.0
wsproto==1.3.1

## YAML / serialization
PyYAML==6.0.3

## Tokens/markdown extras
MarkupSafe==2.1.5
setuptools==80.9.0
tqdm==4.67.1
wheel==0.45.1
